Database

Description

Module containing functions that operate on the database urls.db

Documentation

dbConnection :: IO Connection

Method to create the database urls.db It uses connectSqlite3 from Database.HDBC.Sqlite3

urlNotInDB :: Connection -> URL -> IO Bool

storeURLs

Arguments

:: Connection

-> [URL]
List of URLs to be stored on the database

-> IO ()

Method to store a list of urls into the database (table url)

printURLs :: Connection -> IO ()

Method to display all the URLs on the database. It uses getURLs.

getURLs :: Connection -> IO [URL]

Method to retrieve all the URLs on the database.

getUnprocessedURLs :: Connection -> IO [URL]

Method to retrieve all the URLs on the database.

unfoldDB :: Connection -> IO ()

Method to retrive all the URLs from the database, parse and extract new links from these, and store those links on the database. This method could be made more efficient by remembering which URLs have already been processed, so that next time unfoldDB is called only newly retrieved urls should be processed.

process

Arguments

:: Connection

-> [URL]
List of URLs to be processed

-> IO ()

Method to donwload a given list of URLs, extract new links and store these on the database.

Orphan instances

FromRow Entry

Instance details

Methods

fromRow :: RowParser Entry

ToRow Entry

Instance details

Methods

toRow :: Entry -> [SQLData]