Fetch

Description

Module that performs http requests as well as parsing returned html.

Documentation

type URL = String

type HTML = String

downloadURL :: URL -> IO String

parseURLs

Arguments

:: HTML
String representation of html file

-> [URL]

Basic method for extracting all urls of a given text (String)

isHTML :: URL -> Bool

getHTMLpages :: HTML -> [URL]